requirements.txt
boto3==1.9.117
botocore==1.12.117
contextlib2==0.5.5
docutils==0.14
jmespath==0.9.4
python-dateutil==2.8.0
PyYAML==5.1
s3transfer==0.2.0
schema==0.7.0
six==1.12.0
urllib3==1.24.1
